pyproject.toml
[project]
name = "rfp-analyzer"
version = "0.1.0"
description = "RFP Analyzer using Azure Content Understanding and Microsoft Agent Framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.52.2",
    "agent-framework",
    "azure-identity>=1.19.0",
    "azure-ai-documentintelligence>=1.0.2",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "openai>=2.15.0",
    "requests>=2.32.5",
    "azure-monitor-opentelemetry>=1.8.3",
]

[tool.uv]
prerelease = "allow"

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
]
